+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.32.6](https://github.com/cameronsjo/bosun/compare/v0.32.5...v0.32.6) (2026-03-24)
+
+
+### Bug Fixes
+
+* **reconcile:** address CodeRabbit review feedback for removeStaleFiles ([94c3bb3](https://github.com/cameronsjo/bosun/commit/94c3bb32d4d671bdfb18e8b79aefa2ca382c2451))
+* **reconcile:** log warnings when removeStaleFiles cannot delete files ([f39e4af](https://github.com/cameronsjo/bosun/commit/f39e4af836f84e3dc6d1dc7fac270321deccf117))
+
+## [0.32.5](https://github.com/cameronsjo/bosun/compare/v0.32.4...v0.32.5) (2026-03-24)
+
+
+### Bug Fixes
+
+* **reconcile:** fire all post-sync hooks unconditionally for remote deploys ([1fe3b29](https://github.com/cameronsjo/bosun/commit/1fe3b29d6dd94015338aa753dc36cf51696094bf))
+* **reconcile:** fire all post-sync hooks unconditionally for remote deploys ([54d22eb](https://github.com/cameronsjo/bosun/commit/54d22ebc61e76676be7f42ee1d7dde9bc6d2f4eb)), closes [#197](https://github.com/cameronsjo/bosun/issues/197)
+
 ## [0.32.4](https://github.com/cameronsjo/bosun/compare/v0.32.3...v0.32.4) (2026-03-23)